refactor(ai-tools): extract getAvailableRoles to operation-handlers/resource-operations
